refactor: Changed the ErrorHandler from an interface to a concrete singleton
