+++ app/src/main/java/com/carbondev/carboncheck/data/remote/repository/UserRepositoryRemoteImplementation.kt
@@ -1,7 +1,7 @@
 package com.carbondev.carboncheck.data.remote.repository
 
 import com.carbondev.carboncheck.data.remote.supabase.UserRemoteDataSource
-import com.carbondev.carboncheck.domain.error.ErrorHandler
+import com.carbondev.carboncheck.domain.exception.ErrorHandler
 import com.carbondev.carboncheck.domain.common.ErrorType
 import com.carbondev.carboncheck.domain.common.Result
 import com.carbondev.carboncheck.domain.model.User
