+++ app/src/main/java/com/carbondev/carboncheck/data/remote/repository/VendorRepositoryRemoteImplementation.kt
@@ -3,7 +3,7 @@ package com.carbondev.carboncheck.data.remote.repository
 import com.carbondev.carboncheck.data.remote.supabase.VendorRemoteDataSource
 import com.carbondev.carboncheck.domain.common.ErrorType
 import com.carbondev.carboncheck.domain.common.Result
-import com.carbondev.carboncheck.domain.error.ErrorHandler
+import com.carbondev.carboncheck.domain.exception.ErrorHandler
 import com.carbondev.carboncheck.domain.model.Vendor
 import com.carbondev.carboncheck.domain.repository.VendorRepository
 import timber.log.Timber
