--- app/src/main/java/com/carbondev/carboncheck/domain/error/ErrorHandler.kt
@@ -1,11 +1,0 @@
-package com.carbondev.carboncheck.domain.error
-
-import com.carbondev.carboncheck.domain.common.ErrorType
-
-/**
- * Interface for handling errors in the domain layer.
- * It maps exceptions to domain-specific error types.
- */
-interface ErrorHandler {
-    fun mapToDomainError(exception: Exception): ErrorType
-}
